Refactor script validation logic
