+++ src/qt/addresstablemodel.cpp
@@ -4,6 +4,7 @@
 
 #include "wallet.h"
 #include "base58.h"
+#include "script/standard.h"
 
 #include <QFont>
 #include <QColor>
@@ -59,7 +60,7 @@ public:
         cachedAddressTable.clear();
         {
             LOCK(wallet->cs_wallet);
-            BOOST_FOREACH(const PAIRTYPE(CTxDestination, std::string)& item, wallet->mapAddressBook)
+            for(const PAIRTYPE(CTxDestination, std::string)& item : wallet->mapAddressBook)
             {
                 const CBitcoinAddress& address = item.first;
                 const std::string& strName = item.second;
