+++ src/qt/transactionrecord.cpp
@@ -7,6 +7,7 @@
 #include "wallet.h"
 #include "base58.h"
 #include "validation.h"
+#include "script/standard.h"
 
 /* Return positive answer if transaction should be shown in list.
  */
